move table to own module
